Add tests for sorted wells.
Removed sample sheet concern and moved behaviour back to models.
