+++ .rubocop.yml
@@ -15,13 +15,10 @@ AllCops:
     # GitHub actions installs gems in the vendor directory, which causes rubocop to try
     # and cop our gems. We don't want it to.
     - vendor/**/*
-    - 'config/initializers/secret_token.rb'
     - 'config/initializers/backtrace_silencers.rb'
     - 'config/environments/*'
     - 'bin/*'
     - 'config/application.rb'
-    - 'config/initializers/content_security_policy.rb'
-    - 'config/initializers/new_framework_defaults_7_0.rb'
 
 # The cop is meant to enforce conventions regarding variables containing numbers
 # However in 0.43.0 it is bugged. Both changed on master and disabled by default
@@ -58,6 +55,11 @@ Metrics/MethodLength:
   Exclude:
     - 'db/migrate/*'
     - 'config/initializers/01_proc.rb'
+
+Metrics/ClassLength:
+  Max: 100
+  Exclude:
+    - 'app/models/pacbio/well.rb'
 
 Layout/FirstArrayElementIndentation:
   EnforcedStyle: consistent
